Adapt interception to consolidated V1 utilities
